The icons move depending of what is chosen (desc)
It only shows what is requested when originally created, and moves such to even out the icons and text. This applies mostly to the left side, as the right side is mostly independent from the left side. There is a good likelihood  that none of this code is readable or is badly written because of how tired I am.