Create an overlay for text to be displayed in MilbiG2
Deletes all of the trees that overlap the overlay then creates the overlay then readds the trees back so then they are behind the overlay. Creates two text objects for the title and the body. then can be ended with self.(OverlayObject).complete that will delete itself and the text